fix: move steps files inside new steps folder
