+++ controllers/admin/self-managed/RestorePageBackupSelectionController.php
@@ -7,8 +7,8 @@ use PrestaShop\Module\AutoUpgrade\Parameters\RestoreConfiguration;
 use PrestaShop\Module\AutoUpgrade\Parameters\UpgradeFileNames;
 use PrestaShop\Module\AutoUpgrade\Router\Routes;
 use PrestaShop\Module\AutoUpgrade\Task\TaskType;
-use PrestaShop\Module\AutoUpgrade\Twig\RestoreSteps;
-use PrestaShop\Module\AutoUpgrade\Twig\Steps;
+use PrestaShop\Module\AutoUpgrade\Twig\Steps\RestoreSteps;
+use PrestaShop\Module\AutoUpgrade\Twig\Steps\Steps;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
 class RestorePageBackupSelectionController extends AbstractPageWithStepController
